+++ app/Http/Controllers/Admin/SportController.php
@@ -10,71 +10,45 @@ use Intervention\Image\Facades\Image;
 use Illuminate\Database\QueryException;
 use Yajra\DataTables\Datatables;
 
-use App\Models\SportType;
-use App\Models\Federation;
+use App\Models\Sport;
 
-class SportTypeController extends Controller {
+class SportController extends Controller {
 	public function index (Request $request) {
-		$defaultFederation = $request->has("federation_id") ? $request->federation_id : null;
-
-		$data = [
-			"default_federation" => $defaultFederation,
-			"federation_name" => ($defaultFederation != null) ? Federation::find($defaultFederation)->name : null
-		];
-
-		return view('admin.sport_type.index', $data);
+		return view('admin.sport.index');
 	}
 
 	public function create (Request $request) {
-		$data = [
-			"default_federation" => $request->has("federation_id") ? $request->federation_id : null,
-			"federations" => Federation::get()
-		];
-
-		return view('admin.sport_type.form', $data);
+		return view('admin.sport.form');
 	}
 
 	public function update (Request $request, $id) {
-		$types = SportType::find($id);
+		$sport = Sport::find($id);
 		$data = [
-			"default_federation" => $request->has("federation_id") ? $request->federation_id : null,
-			"federations" => Federation::get(),
-			"data" => $types
+			"data" => $sport
 		];
 
-		return view('admin.sport_type.form', $data);
+		return view('admin.sport.form', $data);
 	}
 
 	public function detail ($id) {
-		$types = SportType::find($id);
+		$types = Sport::find($id);
 		$data = [ "data" => $types ];
 
-		return view('admin.sport_type.detail', $data);
+		return view('admin.sport.detail', $data);
 	}
 
 	public function store (Request $request) {
 		$isCreate = $request->id == null ? true : false;
 		$validation = [
-			'sport_id' => 'required|uuid',
-			'stream_url' => 'required|url',
-			'federation_id' => 'required|uuid'
+			'name' => 'required|max:255',
+			'image' => 'mimes:jpg,png'
 		];
-		
-		if ($isCreate) {
-			array_push($validation, [
-				'image' => 'required|mimes:jpg,png'
-			]);
-		} else {
-			array_push($validation, [
-				'image' => 'mimes:jpg,png'
-			]);
-		}
 			
 		$validated = $request->validate($validation);
 
 		// Upload Image
 		if ($request->hasFile('image')) {
-			if(!Storage::exists("/public/sport/image")) Storage::makeDirectory("/public/sport/image");
+			if(!Storage::exists("/public/sport_type/image")) Storage::makeDirectory("/public/sport_type/image");
 			$file = $request->file('image');
 
 			$filenameWithExt = $request->file('image')->getClientOriginalName();
@@ -82,8 +56,8 @@ class SportTypeController extends Controller {
 			$extension = $request->file('image')->getClientOriginalExtension();
 
 			$filename = $filename.'_'.time().'.'.$extension;
-			$path = storage_path('app/public/sport/image/').$filename;
-			$file->storeAs('public/sport/image', $filename);
+			$path = storage_path('app/public/sport_type/image/').$filename;
+			$file->storeAs('public/sport_type/image', $filename);
 
 			// Resize image
 			$img = Image::make($file->getRealPath())
@@ -92,40 +66,34 @@ class SportTypeController extends Controller {
 				})
 				->save($path, 80);
 		}
-
+		
 		$types = null;
 		if ($isCreate) {
-			$types = new SportType;
+			$types = new Sport;
 			$types->id = Str::uuid();
 		} else {
-			$types = SportType::find($request->id);
+			$types = Sport::find($request->id);
 			if ($request->hasFile('image')) {
-				$oldPath = storage_path('app/public/sport/image/').$types->image;
-				if (file_exists($oldPath)) {
+				$oldPath = storage_path('app/public/sport_type/image/').$types->image;
+				if (file_exists($oldPath) && !is_dir($oldPath)) {
 					unlink($oldPath);
 				}
 			}
 		}
 		
 		try {
-			$types->sport_id = ucwords($request->sport_id);
+			$types->name = ucwords($request->name);
 			if ($request->hasFile('image')) $types->image = $filename;
-			$types->federation_id = $request->federation_id;
-			$types->stream_url = $request->stream_url;
 			$types->save();
 
-			if ($request->is_default_federation == 1) {
-				return redirect()
-					->route("admin.sport.type.index", ["federation_id" => $request->federation_id])
-					->with('success', 'Data successfully saved');
-			} else {
-				return redirect()
-					->route("admin.sport.type.index")
-					->with('success', 'Data successfully saved');
-			}
+			return redirect()
+				->route("admin.sport.index")
+				->with('success', 'Data successfully saved');
 
 		} catch (QueryException $exception) {
-			if ($request->hasFile('image')) unlink($path);
+			if ($request->hasFile('image')) {
+				unlink($path);
+			}
 			return redirect()->back()
 				->withErrors($exception->getMessage());
 		}
@@ -137,7 +105,7 @@ class SportTypeController extends Controller {
 		]);
 
 		try {
-			$type = SportType::find($request->id);
+			$type = Sport::find($request->id);
 			$type->delete();
 
 			session()->flash('message', "{$type->name} successfully deleted");
@@ -156,16 +124,13 @@ class SportTypeController extends Controller {
 	public function list (Request $request) {
 		$search = $request->has('search') ? $request->search : null;
 		$showAll = $request->has('showall') ? (boolean) $request->showall : false;
-		$federation_id = $request->has('federation_id') ? $request->federation_id : null;
 
 		$page = $request->has('page') ? $request->page : 1;
 		if (empty($page)) $page = 1; 
 		$limit = 10;
 
-		$model = SportType::when($search != null, function ($query) use ($search) {
+		$model = Sport::when($search != null, function ($query) use ($search) {
 			$query->where('name', 'LIKE', '%'.$search.'%');
-		})->when($federation_id != null, function ($query) use ($federation_id) {
-			$query->where("federation_id", $federation_id);
 		});
 
 		$types = clone($model)->when(!$showAll, function ($query) use ($limit, $page) {
@@ -191,17 +156,13 @@ class SportTypeController extends Controller {
 	}
 
 	public function listDatatable(Request $request) {
-		$data = SportType::with(["federation", "sport"])
-			->when($request->federation_id != null, function ($query) use ($request) {
-				$query->where("federation_id", $request->federation_id);
-			})
-			->get();
+		$data = Sport::get();
 		
 		return Datatables::of($data)->make(true);
 	}
 
 	public function validateName (Request $request) {
-		$data = SportType::where('name', $request->name)->first();
+		$data = Sport::where('name', $request->name)->first();
 
 		if ($data) {
 			return response()->json([
